+++ examples/react-fcl-demo/src/components/Card.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Card = styled.div`
+  margin: 10px 5px;
+  padding: 10px;
+  border: 1px solid #c0c0c0;
+  border-radius: 5px;
+`
+
+export default Card