+++ examples/react-fcl-demo/src/components/Section.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Section = styled.div`
+  margin: 15px;
+  padding: 10px;
+  border: 1px solid #c0c0c0;
+  border-radius: 5px;
+`
+
+export default Section