+++ examples/react-fcl-demo/src/model/Point.ts
@@ -1,0 +1,11 @@
+class Point {
+  public x: number;
+  public y: number;
+
+  constructor (p: Point) {
+    this.x = p.x
+    this.y = p.y
+  }
+}
+
+export default Point;