+++ packages/fcl/src/fcl.js
@@ -42,4 +42,4 @@ export {
   ref,
 } from "@onflow/sdk"
 
-export const VERSION = "0.0.56"
+export const VERSION = "0.0.60"