+++ packages/fcl/src/resolve/index.js
@@ -1,23 +1,17 @@
-import {
-  resolve as sdkResolve,
-  resolveParams,
-  resolveArguments,
-  resolveAccounts,
-  resolveSignatures,
-  resolveValidators,
-  resolveRefBlockId,
-  // resolveProposerSequenceNumber,
-} from "@onflow/sdk"
+import { resolveRefBlockId } from "@onflow/sdk"
+import {pipe} from "@onflow/interaction"
+import {resolveCadence} from "@onflow/sdk-resolve-cadence"
+import {resolveArguments} from "@onflow/sdk-resolve-arguments"
+import {resolveAccounts} from "@onflow/sdk-resolve-accounts"
+import {resolveSignatures} from "@onflow/sdk-resolve-signatures"
+import {resolveValidators} from "@onflow/sdk-resolve-validators"
 import {config} from "@onflow/config"
 
-export const resolve = async ix => {
-  return sdkResolve(ix, [
-    resolveRefBlockId({node: await config().get("accessNode.api")}),
-    // resolveProposerSequenceNumber({ node: await config().get("accessNode.api") }),
-    resolveParams,
-    resolveArguments,
-    resolveAccounts,
-    resolveSignatures,
-    resolveValidators,
-  ])
-}
+export const resolve = async ix => pipe(ix, [
+  resolveCadence,
+  resolveArguments,
+  resolveAccounts,
+  resolveRefBlockId({node: await config().get("accessNode.api")}),
+  resolveValidators,
+  resolveSignatures,
+])