+++ packages/sdk-build-get-account/src/index.js
@@ -1,0 +1,11 @@
+import {pipe, makeGetAccount, Ok} from "@onflow/interaction"
+
+export function getAccount(addr) {
+  return pipe([
+    makeGetAccount,
+    ix => {
+      ix.accountAddr = addr
+      return Ok(ix)
+    }
+  ])
+}