+++ packages/sdk-build-get-block-by-id/src/index.js
@@ -1,0 +1,11 @@
+import {pipe, Ok, makeGetBlockById} from "@onflow/interaction"
+
+export function getBlockById(id) {
+  return pipe([
+    makeGetBlockById,
+    ix => {
+      ix.block.id = id
+      return Ok(ix)
+    }
+  ])
+}