+++ packages/sdk/src/resolve/resolve-params.js
@@ -1,37 +1,18 @@
-import {isTransaction, isScript, get, Ok} from "@onflow/interaction"
+import {resolveCadence} from "@onflow/sdk-resolve-cadence"
 
-const isFn = d => typeof d === "function"
-const isString = s => typeof s === "string"
+export function resolveParams (ix) {
+  console.error(
+      `
+      %cFCL/SDK Deprecation Notice
+      ============================
 
-export const resolveParams = async (ix) => {
-  if (!(isTransaction(ix) || isScript(ix))) return Ok(ix)
-  const cadence = get(ix, 'ix.cadence')
-  if (isString(cadence)) {
-    ix.message.cadence = cadence
-    return Ok(ix)
-  }
-  if (isFn(cadence)) {
-    let unresolvedParams = Object
-      .values(ix.params)
-    let params = await Promise.all(unresolvedParams.map(
-      async function resParam(up) {
-        if (typeof up.resolve === "function") return ({
-          ...await up.resolve(),
-          tempId: up.tempId
-        })
-        return up
-      }
-    ))
-    params.forEach(p => {
-      ix.params[p.tempId] = p
-    })
-    params = Object.fromEntries(params
-      .filter(param => param.key != null)
-      .map(param => [param.key, param.xform.asInjection(param.value)]))
+      The resolver sdk.resolveParams is being replaced with the package @onflow/sdk-resolve-cadence
+      You can learn more (including a guide on common transition paths) here: https://github.com/onflow/flow-js-sdk/blob/master/packages/sdk/TRANSITIONS.md#0002-deprecate-resolve-params-resolver
 
-    ix.message.cadence = cadence(params)
+      ============================
+    `,
+      "font-weight:bold;font-family:monospace;"
+    )
 
-    return Ok(ix)
-  }
-  throw new Error("Invalid Cadence Value")
+  return resolveCadence(ix)
 }