+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
+### 0.0.35 -- 2020-09-29
+
+- 2020-09-29 -- Proxy @onflow/sdk-resolve-validators at the top level
+- 2020-09-29 -- Proxy @onflow/sdk-resolve-signatures at the top level
+- 2020-09-29 -- Proxy @onflow/sdk-resolve-accounts at the top level
+- 2020-09-29 -- Proxy @onflow/sdk-resolve-arguments at the top level
+- 2020-09-29 -- VSN `@onflow/send` 0.0.21 -> 0.0.22
+
+### 0.0.34 - 2020-09-29
+
+- 2020-09-29 -- VSN `@onflow/send` 0.0.20 -> 0.0.21
+
+### 0.0.33 - 2020-09-29
+
+- 2020-09-29 -- Proxy sdk.resolveArguments to `@onflow/sdk-resolve-arguments`
+- 2020-09-29 -- VSN `@onflow/sdk-resolve-cadence` 0.0.0 -> 0.0.1
+
+### 0.0.32 -- 2020-09-29
+
+- 2020-09-29 -- Deprecate resolverParams
+
 ### 0.0.31 -- 2020-08-25
 
 - 2020-08-25 -- VSN `@onflow/send` 0.0.17 -> 0.0.20